:truck: Modify : 폴더 구조 변경 components -> template -> pages
폴더 구조 변경 기존 pages와 template에 잘못 들어가있는 부분 수정
