--- src/template/InitPage.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-import SetLogin from "../pages/SetLogin"
-
-export default function InitPage() {
-  return (
-    <SetLogin/>
-  )
-}
-
