+++ src/template/signUp/signUpStyle.js
